Extract test/support/db_helper
