+++ test/helper.rb
@@ -13,6 +13,7 @@ $LOAD_PATH.unshift(__dir__)
 require 'active_support'
 require 'active_record_shards'
 require 'mysql2'
+require 'support/db_helper'
 require 'support/tcp_proxy'
 require 'logger'
 
@@ -195,74 +196,4 @@ module RailsEnvSwitch
   end
 end
 
-module DbHelper
-  class << self
-    def client
-      @client ||= begin
-        config = URI(ENV["MYSQL_URL"] || "mysql://root@127.0.0.1:3306")
-        Mysql2::Client.new(
-          username: config.user,
-          password: config.password,
-          host: config.host,
-          port: config.port
-        )
-      end
-    end
-
-    def each_database(&_block)
-      Dir.glob("test/schemas/*.sql").each do |schema_path|
-        database_name = File.basename(schema_path, ".sql").to_s
-        yield(database_name, schema_path)
-      end
-    end
-
-    def mysql(commands)
-      commands.split(/\s*;\s*/).reject(&:empty?).each do |command|
-        client.query(command)
-      end
-    end
-
-    def drop_databases
-      each_database do |database_name, _schema_path|
-        DbHelper.mysql("DROP DATABASE IF EXISTS #{database_name}")
-      end
-    end
-
-    def create_databases
-      each_database do |database_name, _schema_path|
-        DbHelper.mysql("CREATE DATABASE IF NOT EXISTS #{database_name}")
-      end
-    end
-
-    def load_database_schemas
-      each_database do |database_name, schema_path|
-        DbHelper.mysql("USE #{database_name}")
-        DbHelper.mysql(File.read(schema_path))
-      end
-    end
-
-    def load_database_configuration(path = 'test/database.yml')
-      # Load the database configuration into ActiveRecord
-      erb_config = IO.read(path)
-      yaml_config = ERB.new(erb_config).result
-      ActiveRecord::Base.configurations = YAML.load(yaml_config) # rubocop:disable Security/YAMLLoad
-    end
-  end
-
-  # Create all databases and then tear them down after test
-  def with_fresh_databases
-    before do
-      DbHelper.drop_databases
-      DbHelper.create_databases
-      DbHelper.load_database_schemas
-
-      clear_global_connection_handler_state
-      DbHelper.load_database_configuration
-    end
-
-    after do
-      DbHelper.drop_databases
-    end
-  end
-end
 Minitest::Spec.extend(DbHelper)
